Refactor Cards component structure for improved organization
